remove interface for service
