+++ src/test/java/com/guysfromusa/carsgame/services/CarServiceTest.java
@@ -3,76 +3,104 @@ package com.guysfromusa.carsgame.services;
 import com.guysfromusa.carsgame.config.SpringContextConfiguration;
 import com.guysfromusa.carsgame.entities.CarEntity;
 import com.guysfromusa.carsgame.entities.enums.CarType;
+import com.guysfromusa.carsgame.repositories.BaseRepositoryTest;
 import com.guysfromusa.carsgame.repositories.CarRepository;
+import org.apache.commons.collections.iterators.SingletonListIterator;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.ArgumentMatcher;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.runners.MockitoJUnitRunner;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import javax.inject.Inject;
 import java.util.List;
+import java.util.Objects;
 
+import static com.guysfromusa.carsgame.entities.enums.CarType.MONSTER;
+import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.argThat;
+import static org.mockito.Mockito.when;
+import static org.springframework.mock.staticmock.AnnotationDrivenStaticEntityMockingControl.verify;
 
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = SpringContextConfiguration.class)
+@RunWith(MockitoJUnitRunner.class)
 public class CarServiceTest {
 
-    @Inject
+    @Mock
     private CarRepository carRepository;
 
-    @Inject
+    @InjectMocks
     private CarService carService;
 
     @Test
     public void shouldAddNewCar(){
         //given
-        CarType monsterCarType = CarType.MONSTER;
+        CarType monsterCarType = MONSTER;
         String carName = "My-First-CarEntity-Added";
 
         //when
-        Long addedCarId = addCar(monsterCarType, carName);
+        addCar(monsterCarType, carName);
 
         //then
-        assertThat(addedCarId).isGreaterThan(0);
+        Mockito.verify(carRepository).save(argThat(new ArgumentMatcher<CarEntity>() {
+            @Override
+            public boolean matches(Object argument) {
+                CarEntity carEntity = (CarEntity) argument;
+                return Objects.equals(carEntity.getName(), carName) && Objects.equals(carEntity.getCarType(), MONSTER);
+            }
+        }));
     }
 
     @Test
     public void shouldDeleteCar(){
         //given
-        String carName = "My-First-To-Be-Deleted";
-        CarType monsterCarType = CarType.MONSTER;
+        String givenCarName = "My-First-To-Be-Deleted";
 
         //when
-        addCar(monsterCarType, carName);
-        long deletedCarId = carService.deleteCarByName(carName);
+        carService.deleteCarByName(givenCarName);
 
         //then
-        assertTrue(deletedCarId > 0);
+        Mockito.verify(carRepository).deleteByName(argThat(new ArgumentMatcher<String>() {
+            @Override
+            public boolean matches(Object argument) {
+                String carName = (String) argument;
+                return Objects.equals(carName, givenCarName);
+            }
+        }));
     }
 
     @Test
     public void shouldLoadAtLeastOneCar(){
         //given
         String carName = "My-Second-Car";
-        CarType monsterCarType = CarType.MONSTER;
+        List<CarEntity> cars = singletonList(new CarEntity() {{
+            this.setName(carName);
+        }});
+        when(carRepository.findAll()).thenReturn(cars);
 
         //when
-        addCar(monsterCarType, carName);
         Iterable<CarEntity> carEntities = carService.loadAllCars();
 
         //then
-        assertThat(carEntities).isNotEmpty();
+        assertThat(carEntities).isNotEmpty().extracting(CarEntity::getName).contains(carName);
 
     }
+
 
     private Long addCar(CarType carType, String carName){
         return carService.addCar(carType, carName);
     }
+
+
 
 }
